scripts/requirements.txt
python-dotenv==1.0.0
psycopg2==2.9.9
Faker==22.2.0